+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "iwdd.net-nextjs",
-  "version": "0.1.0",
+  "version": "0.2.0",
   "engines": {
     "node": "20"
   },
@@ -8,7 +8,6 @@
   "scripts": {
     "dev": "next dev",
     "build": "next build",
-    "postbuild": "next-sitemap",
     "start": "next start",
     "lint": "next lint"
   },
@@ -16,7 +15,6 @@
     "dayjs": "1.11.11",
     "js-yaml": "4.1.0",
     "next": "14.2.3",
-    "next-sitemap": "4.2.3",
     "node-html-parser": "6.1.13",
     "react": "18.3.1",
     "react-dom": "18.2.0"
