+++ src/app/layout.tsx
@@ -1,0 +1,25 @@
+import '@/styles/globals.css'
+
+import { Shippori_Mincho } from 'next/font/google'
+import { type ReactNode } from 'react'
+
+const shipporiMincho = Shippori_Mincho({
+  display: 'swap',
+  subsets: ['latin'],
+  variable: '--font-shipporiMincho',
+  weight: ['400', '700'],
+})
+
+export default function RootLayout({
+  children,
+}: Readonly<{
+  children: ReactNode
+}>) {
+  return (
+    <html lang="en">
+      <body className={`${shipporiMincho.variable} font-shipporiMincho`}>
+        {children}
+      </body>
+    </html>
+  )
+}
